+++ scheduler/conf.c
@@ -3589,6 +3589,7 @@ read_cups_files_conf(cups_file_t *fp)	/* I - File to read from */
   static const char * const prohibited_env[] =
   {					/* Prohibited environment variables */
     "APPLE_LANGUAGE",
+    "AUTH_BEARER",
     "AUTH_DOMAIN",
     "AUTH_INFO_REQUIRED",
     "AUTH_NEGOTIATE",